chore: renaming operator
