+++ src/core/operators/md5_hash.py
