+++ src/core/operators/test_md5_hash.py
@@ -1,12 +1,12 @@
 import unittest
 from unittest.case import skip
-import md5
+import core.operators.md5_hash as md5_hash
 
 class Test(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         # initialize operator
-        md5.initialize(param={})
+        md5_hash.initialize(param={})
 
     @classmethod
     def tearDownClass(cls):
@@ -15,6 +15,6 @@ class Test(unittest.TestCase):
 
     def test_sample_media_from_disk(self):
         media_file_path = r'sample_data/sample-cat-video.mp4'
-        md5_hash = md5.run(media_file_path)
+        md5_hash = md5_hash.run(media_file_path)
         print(md5_hash)
         self.assertEqual(32, len(md5_hash))
